refactor: dto 위치 리팩터링
