+++ src/main/java/subway/application/StationService.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import subway.domain.station.Station;
 import subway.domain.station.StationName;
-import subway.dto.StationRequest;
-import subway.dto.StationResponse;
+import subway.dto.request.StationRequest;
+import subway.dto.response.StationResponse;
 import subway.repository.StationRepository;
 
 @Transactional(readOnly = true)
