+++ src/main/java/subway/dto/response/StationResponse.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.dto.response;
 
 public class StationResponse {
     private Long id;
